start working on cloud commands
