+++ plugin/src/main/java/org/screamingsandals/bedwars/Main.java
@@ -1,7 +1,7 @@
 package org.screamingsandals.bedwars;
 
-import lombok.Getter;
 import org.bukkit.event.Listener;
+import org.screamingsandals.bedwars.commands.CommandGenerator;
 import org.screamingsandals.bedwars.config.MainConfig;
 import org.screamingsandals.bedwars.config.RecordSave;
 import org.screamingsandals.bedwars.lib.lang.I18n;
@@ -11,7 +11,6 @@ import org.bstats.bukkit.Metrics;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.Material;
-import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
@@ -23,7 +22,6 @@ import org.screamingsandals.bedwars.api.game.GameStatus;
 import org.screamingsandals.bedwars.game.GameStore;
 import org.screamingsandals.bedwars.api.statistics.PlayerStatisticsManager;
 import org.screamingsandals.bedwars.api.utils.ColorChanger;
-import org.screamingsandals.bedwars.commands.*;
 import org.screamingsandals.bedwars.database.DatabaseManager;
 import org.screamingsandals.bedwars.game.Game;
 import org.screamingsandals.bedwars.game.GamePlayer;
@@ -45,10 +43,13 @@ import org.screamingsandals.bedwars.lib.nms.holograms.HologramManager;
 import org.screamingsandals.bedwars.lib.nms.utils.ClassStorage;
 import org.screamingsandals.bedwars.lib.signmanager.SignListener;
 import org.screamingsandals.bedwars.lib.signmanager.SignManager;
+import org.screamingsandals.lib.command.CloudConstructor;
 import org.screamingsandals.lib.event.EventManager;
 import org.screamingsandals.lib.material.MaterialMapping;
+import org.screamingsandals.lib.player.PlayerMapper;
 import org.screamingsandals.lib.plugin.PluginContainer;
 import org.screamingsandals.lib.plugin.PluginManager;
+import org.screamingsandals.lib.sender.CommandSenderWrapper;
 import org.screamingsandals.lib.utils.ControllableImpl;
 import org.screamingsandals.lib.utils.InitUtils;
 import org.screamingsandals.lib.utils.PlatformType;
@@ -61,10 +62,8 @@ import org.spongepowered.configurate.serialize.SerializationException;
 import org.spongepowered.configurate.yaml.YamlConfigurationLoader;
 import pronze.lib.scoreboards.ScoreboardManager;
 
-import java.io.BufferedReader;
 import java.io.File;
 import java.io.IOException;
-import java.io.InputStreamReader;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -95,13 +94,13 @@ import static org.screamingsandals.bedwars.lib.lang.I18n.i18n;
 })
 @Init(services = {
         SimpleInventoriesCore.class,
-        EventManager.class
+        EventManager.class,
+        CloudConstructor.class,
+        CommandGenerator.class
 })
 public class Main extends PluginContainer implements BedwarsAPI {
-    @Getter
-    private static final String buildInfo = VersionInfo.BUILD_NUMBER;
-
     private static Main instance;
+
     private String version;
     private boolean isDisabling = false;
     private boolean isSpigot, isLegacy;
@@ -118,7 +117,6 @@ public class Main extends PluginContainer implements BedwarsAPI {
     private DatabaseManager databaseManager;
     private PlayerStatisticManager playerStatisticsManager;
     private StatisticsHolograms hologramInteraction;
-    private HashMap<String, BaseCommand> commands;
     private ColorChanger colorChanger;
     private SignManager signManager;
     private HologramManager manager;
@@ -244,9 +242,9 @@ public class Main extends PluginContainer implements BedwarsAPI {
         return instance.playersInGame.containsKey(player);
     }
 
-    public static void sendGameListInfo(CommandSender player) {
+    public static void sendGameListInfo(CommandSenderWrapper senderWrapper) {
         for (Game game : instance.games.values()) {
-            player.sendMessage((game.getStatus() == GameStatus.DISABLED ? "§c" : "§a") + game.getName() + "§f "
+            senderWrapper.sendMessage((game.getStatus() == GameStatus.DISABLED ? "§c" : "§a") + game.getName() + "§f "
                     + game.countPlayers());
         }
     }
@@ -354,10 +352,6 @@ public class Main extends PluginContainer implements BedwarsAPI {
         return instance.hologramInteraction;
     }
 
-    public static HashMap<String, BaseCommand> getCommands() {
-        return instance.commands;
-    }
-
     public static List<Entity> getGameEntities(Game game) {
         List<Entity> entityList = new ArrayList<>();
         for (Map.Entry<Entity, Game> entry : instance.entitiesInGame.entrySet()) {
@@ -442,7 +436,7 @@ public class Main extends PluginContainer implements BedwarsAPI {
         }
 
         try {
-            if (Bukkit.getPluginManager().getPlugin("PlaceholderAPI") != null) {
+            if (PluginManager.isEnabled(PluginManager.createKey("PlaceholderAPI").orElseThrow())) {
                 new BedwarsExpansion().register();
             }
         } catch (Throwable ignored) {
@@ -469,30 +463,6 @@ public class Main extends PluginContainer implements BedwarsAPI {
                 partiesEnabled = true;
             }
         }
-
-        commands = new HashMap<>();
-        new AddholoCommand();
-        new AdminCommand();
-        new AutojoinCommand();
-        new AllJoinCommand();
-        new HelpCommand();
-        new JoinCommand();
-        new LeaveCommand();
-        new ListCommand();
-        new RejoinCommand();
-        new ReloadCommand();
-        new RemoveholoCommand();
-        new StatsCommand();
-        new MainlobbyCommand();
-        new LeaderboardCommand();
-        new DumpCommand();
-        new LanguageCommand();
-        if (partiesEnabled)
-            new PartyCommand();
-
-        final var cmd = new BwCommandsExecutor();
-        getCommand("bw").setExecutor(cmd);
-        getCommand("bw").setTabCompleter(cmd);
 
         registerBedwarsListener(new PlayerListener());
         if (versionNumber >= 109) {
@@ -565,21 +535,21 @@ public class Main extends PluginContainer implements BedwarsAPI {
             Bukkit.getMessenger().registerOutgoingPluginChannel(this.getPluginDescription().as(JavaPlugin.class), "BungeeCord");
         }
 
-        Bukkit.getConsoleSender().sendMessage(ChatColor.AQUA + "============" + ChatColor.RED + "===" + ChatColor.WHITE + "======  by ScreamingSandals <Misat11, Ceph, Pronze>");
-        Bukkit.getConsoleSender()
+        PlayerMapper.getConsoleSender().sendMessage(ChatColor.AQUA + "============" + ChatColor.RED + "===" + ChatColor.WHITE + "======  by ScreamingSandals <Misat11, Ceph, Pronze>");
+        PlayerMapper.getConsoleSender()
                 .sendMessage(ChatColor.AQUA + "+ Screaming " + ChatColor.RED + "Bed" + ChatColor.WHITE + "Wars +  " + ChatColor.GOLD + "Version: " + version + " " + (PremiumBedwars.isPremium() ? ChatColor.AQUA + "PREMIUM" : ChatColor.GREEN + "FREE"));
-        Bukkit.getConsoleSender()
-                .sendMessage(ChatColor.AQUA + "============" + ChatColor.RED + "===" + ChatColor.WHITE + "======  " + (snapshot ? ChatColor.RED + "SNAPSHOT VERSION (" + buildInfo + ") - Use at your own risk" : ChatColor.GREEN + "STABLE VERSION"));
+        PlayerMapper.getConsoleSender()
+                .sendMessage(ChatColor.AQUA + "============" + ChatColor.RED + "===" + ChatColor.WHITE + "======  " + (snapshot ? ChatColor.RED + "SNAPSHOT VERSION (" + VersionInfo.BUILD_NUMBER + ") - Use at your own risk" : ChatColor.GREEN + "STABLE VERSION"));
         if (isVault) {
-            Bukkit.getConsoleSender().sendMessage(ChatColor.RED + "[B" + ChatColor.WHITE + "W] " + ChatColor.GOLD + "Found Vault");
+            PlayerMapper.getConsoleSender().sendMessage(ChatColor.RED + "[B" + ChatColor.WHITE + "W] " + ChatColor.GOLD + "Found Vault");
         }
         if (!isSpigot) {
-            Bukkit.getConsoleSender()
+            PlayerMapper.getConsoleSender()
                     .sendMessage(ChatColor.RED + "[B" + ChatColor.WHITE + "W] " + ChatColor.RED + "WARNING: You are not using Spigot. Some features may not work properly.");
         }
 
         if (versionNumber < 109) {
-            Bukkit.getConsoleSender().sendMessage(
+            PlayerMapper.getConsoleSender().sendMessage(
                     ChatColor.RED + "[B" + ChatColor.WHITE + "W] " + ChatColor.RED + "IMPORTANT WARNING: You are using version older than 1.9! This version is not officially supported, and some features may not work at all!");
         }
 
@@ -646,10 +616,10 @@ public class Main extends PluginContainer implements BedwarsAPI {
         final var pluginId = 7147;
         metrics = new Metrics(this.getPluginDescription().as(JavaPlugin.class), pluginId);
         metrics.addCustomChart(new Metrics.SimplePie("edition", () -> PremiumBedwars.isPremium() ? "Premium" : "Free"));
-        metrics.addCustomChart(new Metrics.SimplePie("build_number", () -> buildInfo));
+        metrics.addCustomChart(new Metrics.SimplePie("build_number", () -> VersionInfo.BUILD_NUMBER));
 
-        Bukkit.getConsoleSender().sendMessage(ChatColor.WHITE + "Everything is loaded! If you like our work, consider visiting our Patreon! <3");
-        Bukkit.getConsoleSender().sendMessage(ChatColor.WHITE + "https://www.patreon.com/screamingsandals");
+        PlayerMapper.getConsoleSender().sendMessage(ChatColor.WHITE + "Everything is loaded! If you like our work, consider visiting our Patreon! <3");
+        PlayerMapper.getConsoleSender().sendMessage(ChatColor.WHITE + "https://www.patreon.com/screamingsandals");
     }
 
     public void onDisable() {
