+++ plugin/src/main/java/org/screamingsandals/bedwars/commands/AutojoinCommand.java
@@ -1,40 +1,37 @@
 package org.screamingsandals.bedwars.commands;
 
-import org.screamingsandals.bedwars.Main;
-import org.screamingsandals.bedwars.api.game.Game;
-import org.bukkit.command.CommandSender;
+import cloud.commandframework.Command;
+import cloud.commandframework.CommandManager;
 import org.bukkit.entity.Player;
+import org.screamingsandals.bedwars.Main;
+import org.screamingsandals.lib.sender.CommandSenderWrapper;
 
-import java.util.List;
-
-import static org.screamingsandals.bedwars.lib.lang.I18n.i18n;
+import static org.screamingsandals.bedwars.lib.lang.I.i18n;
 
 public class AutojoinCommand extends BaseCommand {
-
-    public AutojoinCommand() {
-        super("autojoin", AUTOJOIN_PERMISSION, false, Main.getConfigurator().node("default-permissions", "autojoin").getBoolean());
+    public AutojoinCommand(CommandManager<CommandSenderWrapper> manager) {
+        super(manager, "autojoin", BedWarsPermission.AUTOJOIN_PERMISSION, false);
     }
 
     @Override
-    public boolean execute(CommandSender sender, List<String> args) {
-        Player player = (Player) sender;
-        if (Main.isPlayerInGame(player)) {
-            player.sendMessage(i18n("you_are_already_in_some_game"));
-            return true;
-        }
+    protected void construct(Command.Builder<CommandSenderWrapper> commandSenderWrapperBuilder) {
+        manager.command(
+                commandSenderWrapperBuilder
+                    .handler(commandContext -> {
+                        // TODO: Use Wrapper (bedwars changes needed)
+                        var player = commandContext.getSender().as(Player.class);
+                        if (Main.isPlayerInGame(player)) {
+                            player.sendMessage(i18n("you_are_already_in_some_game"));
+                            return;
+                        }
 
-        Game game = Main.getInstance().getFirstWaitingGame();
-        if (game == null) {
-            player.sendMessage(i18n("there_is_no_empty_game"));
-        } else {
-            game.joinToGame(player);
-        }
-        return true;
+                        var game = Main.getInstance().getFirstWaitingGame();
+                        if (game == null) {
+                            player.sendMessage(i18n("there_is_no_empty_game"));
+                        } else {
+                            game.joinToGame(player);
+                        }
+                    })
+        );
     }
-
-    @Override
-    public void completeTab(List<String> completion, CommandSender sender, List<String> args) {
-        // Nothing to add.
-    }
-
 }
