+++ plugin/src/main/java/org/screamingsandals/bedwars/commands/HelpCommand.java
@@ -1,116 +1,114 @@
 package org.screamingsandals.bedwars.commands;
 
+import cloud.commandframework.Command;
+import cloud.commandframework.CommandManager;
 import org.screamingsandals.bedwars.Main;
-import org.bukkit.command.CommandSender;
-import org.bukkit.command.ConsoleCommandSender;
-import org.bukkit.entity.Player;
+import org.screamingsandals.lib.sender.CommandSenderWrapper;
 
-import java.util.Collections;
-import java.util.List;
-
-import static org.screamingsandals.bedwars.lib.lang.I18n.i18nonly;
+import static org.screamingsandals.bedwars.lib.lang.I.i18nonly;
 
 public class HelpCommand extends BaseCommand {
-
-    public HelpCommand() {
-        super("help", Collections.EMPTY_LIST, true, true);
+    public HelpCommand(CommandManager<CommandSenderWrapper> manager) {
+        super(manager, "help", null, true);
     }
 
     @Override
-    public boolean execute(CommandSender sender, List<String> args) {
-        if (sender instanceof Player) {
-            sendHelp((Player) sender);
-        } else if (sender instanceof ConsoleCommandSender) {
-            sendConsoleHelp((ConsoleCommandSender) sender);
-        }
-        return true;
+    protected void construct(Command.Builder<CommandSenderWrapper> commandSenderWrapperBuilder) {
+        manager.command(commandSenderWrapperBuilder
+                .handler(commandContext -> {
+                    // TODO: use more generic way
+                    var sender = commandContext.getSender();
+                    if (sender.getType() == CommandSenderWrapper.Type.PLAYER) {
+                        sender.sendMessage(i18nonly("help_title").replace("%version%", Main.getVersion()));
+                        if (sender.hasPermission(BedWarsPermission.JOIN_PERMISSION.asPermission())) {
+                            sender.sendMessage(i18nonly("help_bw_join"));
+                        }
+                        if (sender.hasPermission(BedWarsPermission.LEAVE_PERMISSION.asPermission())) {
+                            sender.sendMessage(i18nonly("help_bw_leave"));
+                        }
+                        if (sender.hasPermission(BedWarsPermission.REJOIN_PERMISSION.asPermission())) {
+                            sender.sendMessage(i18nonly("help_bw_rejoin"));
+                        }
+                        if (sender.hasPermission(BedWarsPermission.AUTOJOIN_PERMISSION.asPermission())) {
+                            sender.sendMessage(i18nonly("help_bw_autojoin"));
+                        }
+                        if (sender.hasPermission(BedWarsPermission.LIST_PERMISSION.asPermission())) {
+                            sender.sendMessage(i18nonly("help_bw_list"));
+                        }
+                        if (sender.hasPermission(BedWarsPermission.LEADERBOARD_PERMISSION.asPermission())) {
+                            sender.sendMessage(i18nonly("help_bw_leaderboard"));
+                        }
+
+                        if ((sender.hasPermission(BedWarsPermission.STATS_PERMISSION.asPermission()))) {
+                            if (sender.hasPermission(BedWarsPermission.ADMIN_PERMISSION.asPermission()) || sender.hasPermission(BedWarsPermission.OTHER_STATS_PERMISSION.asPermission())) {
+                                sender.sendMessage(i18nonly("help_bw_stats_other"));
+                            } else {
+                                sender.sendMessage(i18nonly("help_bw_stats"));
+                            }
+                        }
+
+                        if (sender.hasPermission("bw.admin.alljoin")) {
+                            sender.sendMessage(i18nonly("help_bw_alljoin"));
+                        }
+
+                        if (sender.hasPermission(BedWarsPermission.ADMIN_PERMISSION.asPermission())) {
+                            sender.sendMessage(i18nonly("help_bw_addholo"));
+                            sender.sendMessage(i18nonly("help_bw_removeholo"));
+                            sender.sendMessage(i18nonly("help_bw_mainlobby"));
+
+                            sender.sendMessage(i18nonly("help_bw_admin_info"));
+                            sender.sendMessage(i18nonly("help_bw_admin_add"));
+                            sender.sendMessage(i18nonly("help_bw_admin_lobby"));
+                            sender.sendMessage(i18nonly("help_bw_admin_spec"));
+                            sender.sendMessage(i18nonly("help_bw_admin_pos1"));
+                            sender.sendMessage(i18nonly("help_bw_admin_pos2"));
+                            sender.sendMessage(i18nonly("help_bw_admin_pausecountdown"));
+                            sender.sendMessage(i18nonly("help_bw_admin_post_game_waiting"));
+                            sender.sendMessage(i18nonly("help_bw_admin_customprefix"));
+                            sender.sendMessage(i18nonly("help_bw_admin_minplayers"));
+                            sender.sendMessage(i18nonly("help_bw_admin_time"));
+                            sender.sendMessage(i18nonly("help_bw_admin_team_add"));
+                            sender.sendMessage(i18nonly("help_bw_admin_team_color"));
+                            sender.sendMessage(i18nonly("help_bw_admin_team_maxplayers"));
+                            sender.sendMessage(i18nonly("help_bw_admin_team_spawn"));
+                            sender.sendMessage(i18nonly("help_bw_admin_team_bed"));
+                            sender.sendMessage(i18nonly("help_bw_admin_jointeam"));
+                            sender.sendMessage(i18nonly("help_bw_admin_spawner_add"));
+                            sender.sendMessage(i18nonly("help_bw_admin_spawner_remove"));
+                            sender.sendMessage(i18nonly("help_bw_admin_spawner_reset"));
+                            sender.sendMessage(i18nonly("help_bw_admin_store_add"));
+                            sender.sendMessage(i18nonly("help_bw_admin_store_type"));
+                            sender.sendMessage(i18nonly("help_bw_admin_store_child"));
+                            sender.sendMessage(i18nonly("help_bw_admin_store_adult"));
+                            sender.sendMessage(i18nonly("help_bw_admin_store_remove"));
+                            sender.sendMessage(i18nonly("help_bw_admin_config"));
+                            sender.sendMessage(i18nonly("help_bw_admin_arena_time"));
+                            sender.sendMessage(i18nonly("help_bw_admin_arena_weather"));
+                            sender.sendMessage(i18nonly("help_bw_admin_remove"));
+                            sender.sendMessage(i18nonly("help_bw_admin_edit"));
+                            sender.sendMessage(i18nonly("help_bw_admin_save"));
+                            sender.sendMessage(i18nonly("help_bw_reload"));
+                            sender.sendMessage(i18nonly("help_bw_dump"));
+                        }
+                    } else {
+                        sender.sendMessage(i18nonly("help_title_console").replace("%version%", Main.getVersion()));
+                        sender.sendMessage(i18nonly("help_bw_list"));
+                        sender.sendMessage(i18nonly("help_bw_stats_other"));
+                        sender.sendMessage(i18nonly("help_bw_alljoin"));
+                        sender.sendMessage(i18nonly("help_bw_reload"));
+                    }
+                })
+        );
     }
 
+    /*
+     * Special case, only for help command
+     */
     @Override
-    public void completeTab(List<String> completion, CommandSender sender, List<String> args) {
-        // Nothing to add.
+    public void construct() {
+        var builder = manager.commandBuilder("bw");
+        construct(builder);
+
+        super.construct();
     }
-
-    public void sendConsoleHelp(ConsoleCommandSender console) {
-        console.sendMessage(i18nonly("help_title_console").replace("%version%", Main.getVersion()));
-        console.sendMessage(i18nonly("help_bw_list"));
-        console.sendMessage(i18nonly("help_bw_stats_other"));
-        console.sendMessage(i18nonly("help_bw_alljoin"));
-        console.sendMessage(i18nonly("help_bw_reload"));
-    }
-
-    public void sendHelp(Player player) {
-        player.sendMessage(i18nonly("help_title").replace("%version%", Main.getVersion()));
-        if (hasPermission(player, JOIN_PERMISSION, Main.getConfigurator().node("default-permissions", "join").getBoolean())) {
-            player.sendMessage(i18nonly("help_bw_join"));
-        }
-        if (hasPermission(player, LEAVE_PERMISSION, Main.getConfigurator().node("default-permissions", "leave").getBoolean())) {
-            player.sendMessage(i18nonly("help_bw_leave"));
-        }
-        if (hasPermission(player, REJOIN_PERMISSION, Main.getConfigurator().node("default-permissions", "rejoin").getBoolean())) {
-            player.sendMessage(i18nonly("help_bw_rejoin"));
-        }
-        if (hasPermission(player, AUTOJOIN_PERMISSION, Main.getConfigurator().node("default-permissions", "autojoin").getBoolean())) {
-            player.sendMessage(i18nonly("help_bw_autojoin"));
-        }
-        if (hasPermission(player, LIST_PERMISSION, Main.getConfigurator().node("default-permissions", "list").getBoolean())) {
-            player.sendMessage(i18nonly("help_bw_list"));
-        }
-        if (hasPermission(player, LEADERBOARD_PERMISSION, Main.getConfigurator().node("default-permissions", "leaderboard").getBoolean())) {
-            player.sendMessage(i18nonly("help_bw_leaderboard"));
-        }
-
-        if ((hasPermission(player, STATS_PERMISSION, Main.getConfigurator().node("default-permissions", "stats").getBoolean()))) {
-            if (hasPermission(player, ADMIN_PERMISSION, false) || hasPermission(player, OTHER_STATS_PERMISSION, false)) {
-                player.sendMessage(i18nonly("help_bw_stats_other"));
-            } else {
-                player.sendMessage(i18nonly("help_bw_stats"));
-            }
-        }
-
-        if (player.hasPermission("bw.admin.alljoin")) {
-            player.sendMessage(i18nonly("help_bw_alljoin"));
-        }
-
-        if (hasPermission(player, ADMIN_PERMISSION, false)) {
-            player.sendMessage(i18nonly("help_bw_addholo"));
-            player.sendMessage(i18nonly("help_bw_removeholo"));
-            player.sendMessage(i18nonly("help_bw_mainlobby"));
-
-            player.sendMessage(i18nonly("help_bw_admin_info"));
-            player.sendMessage(i18nonly("help_bw_admin_add"));
-            player.sendMessage(i18nonly("help_bw_admin_lobby"));
-            player.sendMessage(i18nonly("help_bw_admin_spec"));
-            player.sendMessage(i18nonly("help_bw_admin_pos1"));
-            player.sendMessage(i18nonly("help_bw_admin_pos2"));
-            player.sendMessage(i18nonly("help_bw_admin_pausecountdown"));
-            player.sendMessage(i18nonly("help_bw_admin_post_game_waiting"));
-            player.sendMessage(i18nonly("help_bw_admin_customprefix"));
-            player.sendMessage(i18nonly("help_bw_admin_minplayers"));
-            player.sendMessage(i18nonly("help_bw_admin_time"));
-            player.sendMessage(i18nonly("help_bw_admin_team_add"));
-            player.sendMessage(i18nonly("help_bw_admin_team_color"));
-            player.sendMessage(i18nonly("help_bw_admin_team_maxplayers"));
-            player.sendMessage(i18nonly("help_bw_admin_team_spawn"));
-            player.sendMessage(i18nonly("help_bw_admin_team_bed"));
-            player.sendMessage(i18nonly("help_bw_admin_jointeam"));
-            player.sendMessage(i18nonly("help_bw_admin_spawner_add"));
-            player.sendMessage(i18nonly("help_bw_admin_spawner_remove"));
-            player.sendMessage(i18nonly("help_bw_admin_spawner_reset"));
-            player.sendMessage(i18nonly("help_bw_admin_store_add"));
-            player.sendMessage(i18nonly("help_bw_admin_store_type"));
-            player.sendMessage(i18nonly("help_bw_admin_store_child"));
-            player.sendMessage(i18nonly("help_bw_admin_store_adult"));
-            player.sendMessage(i18nonly("help_bw_admin_store_remove"));
-            player.sendMessage(i18nonly("help_bw_admin_config"));
-            player.sendMessage(i18nonly("help_bw_admin_arena_time"));
-            player.sendMessage(i18nonly("help_bw_admin_arena_weather"));
-            player.sendMessage(i18nonly("help_bw_admin_remove"));
-            player.sendMessage(i18nonly("help_bw_admin_edit"));
-            player.sendMessage(i18nonly("help_bw_admin_save"));
-            player.sendMessage(i18nonly("help_bw_reload"));
-            player.sendMessage(i18nonly("help_bw_dump"));
-        }
-    }
-
 }
