+++ plugin/src/main/java/org/screamingsandals/bedwars/commands/ListCommand.java
@@ -1,27 +1,26 @@
 package org.screamingsandals.bedwars.commands;
 
+import cloud.commandframework.Command;
+import cloud.commandframework.CommandManager;
 import org.screamingsandals.bedwars.Main;
-import org.bukkit.command.CommandSender;
+import org.screamingsandals.lib.sender.CommandSenderWrapper;
 
-import java.util.List;
-
-import static org.screamingsandals.bedwars.lib.lang.I18n.i18n;
+import static org.screamingsandals.bedwars.lib.lang.I.i18n;
 
 public class ListCommand extends BaseCommand {
-
-    public ListCommand() {
-        super("list", LIST_PERMISSION, true, Main.getConfigurator().node("default-permissions", "list").getBoolean());
+    public ListCommand(CommandManager<CommandSenderWrapper> manager) {
+        super(manager, "list", BedWarsPermission.LIST_PERMISSION, true);
     }
 
     @Override
-    public boolean execute(CommandSender sender, List<String> args) {
-        sender.sendMessage(i18n("list_header"));
-        Main.sendGameListInfo(sender);
-        return true;
+    protected void construct(Command.Builder<CommandSenderWrapper> commandSenderWrapperBuilder) {
+        manager.command(
+                commandSenderWrapperBuilder
+                    .handler(commandContext -> {
+                        var sender = commandContext.getSender();
+                        sender.sendMessage(i18n("list_header"));
+                        Main.sendGameListInfo(sender);
+                    })
+        );
     }
-
-    @Override
-    public void completeTab(List<String> completion, CommandSender sender, List<String> args) {
-    }
-
 }
