+++ plugin/src/main/java/org/screamingsandals/bedwars/commands/RejoinCommand.java
@@ -1,46 +1,44 @@
 package org.screamingsandals.bedwars.commands;
 
-import org.screamingsandals.bedwars.Main;
-import org.bukkit.command.CommandSender;
+import cloud.commandframework.Command;
+import cloud.commandframework.CommandManager;
 import org.bukkit.entity.Player;
+import org.screamingsandals.bedwars.Main;
+import org.screamingsandals.lib.sender.CommandSenderWrapper;
 
-import java.util.List;
-
-import static org.screamingsandals.bedwars.lib.lang.I18n.i18n;
+import static org.screamingsandals.bedwars.lib.lang.I.i18n;
 
 public class RejoinCommand extends BaseCommand {
-
-    public RejoinCommand() {
-        super("rejoin", REJOIN_PERMISSION, false, Main.getConfigurator().node("default-permissions", "rejoin").getBoolean());
+    public RejoinCommand(CommandManager<CommandSenderWrapper> manager) {
+        super(manager, "rejoin", BedWarsPermission.REJOIN_PERMISSION, false);
     }
 
     @Override
-    public boolean execute(CommandSender sender, List<String> args) {
-        Player player = (Player) sender;
-        if (Main.isPlayerInGame(player)) {
-            player.sendMessage(i18n("you_are_already_in_some_game"));
-            return true;
-        }
+    protected void construct(Command.Builder<CommandSenderWrapper> commandSenderWrapperBuilder) {
+        manager.command(
+                commandSenderWrapperBuilder
+                    .handler(commandContext -> {
+                        // TODO: Use Wrapper (bedwars changes needed)
+                        var player = commandContext.getSender().as(Player.class);
+                        if (Main.isPlayerInGame(player)) {
+                            player.sendMessage(i18n("you_are_already_in_some_game"));
+                            return;
+                        }
 
-        String name = null;
-        if (Main.isPlayerGameProfileRegistered(player)) {
-            name = Main.getPlayerGameProfile(player).getLatestGameName();
-        }
-        if (name == null) {
-            player.sendMessage(i18n("you_are_not_in_game_yet"));
-        } else {
-            if (Main.isGameExists(name)) {
-                Main.getGame(name).joinToGame(player);
-            } else {
-                player.sendMessage(i18n("game_is_gone"));
-            }
-        }
-        return true;
+                        String name = null;
+                        if (Main.isPlayerGameProfileRegistered(player)) {
+                            name = Main.getPlayerGameProfile(player).getLatestGameName();
+                        }
+                        if (name == null) {
+                            player.sendMessage(i18n("you_are_not_in_game_yet"));
+                        } else {
+                            if (Main.isGameExists(name)) {
+                                Main.getGame(name).joinToGame(player);
+                            } else {
+                                player.sendMessage(i18n("game_is_gone"));
+                            }
+                        }
+                    })
+        );
     }
-
-    @Override
-    public void completeTab(List<String> completion, CommandSender sender, List<String> args) {
-        // Nothing to add.
-    }
-
 }
