+++ plugin/src/main/java/org/screamingsandals/bedwars/commands/RemoveholoCommand.java
@@ -1,34 +1,34 @@
 package org.screamingsandals.bedwars.commands;
 
-import org.screamingsandals.bedwars.Main;
-import org.bukkit.command.CommandSender;
+import cloud.commandframework.Command;
+import cloud.commandframework.CommandManager;
 import org.bukkit.entity.Player;
 import org.bukkit.metadata.FixedMetadataValue;
+import org.bukkit.plugin.java.JavaPlugin;
+import org.screamingsandals.bedwars.Main;
+import org.screamingsandals.lib.sender.CommandSenderWrapper;
 
-import java.util.List;
-
-import static org.screamingsandals.bedwars.lib.lang.I18n.i18n;
+import static org.screamingsandals.bedwars.lib.lang.I.i18n;
 
 public class RemoveholoCommand extends BaseCommand {
-
-    public RemoveholoCommand() {
-        super("removeholo", ADMIN_PERMISSION, false, false);
+    public RemoveholoCommand(CommandManager<CommandSenderWrapper> manager) {
+        super(manager, "removeholo", BedWarsPermission.ADMIN_PERMISSION, false);
     }
 
     @Override
-    public boolean execute(CommandSender sender, List<String> args) {
-        Player player = (Player) sender;
-        if (!Main.isHologramsEnabled()) {
-            player.sendMessage(i18n("holo_not_enabled"));
-        } else {
-            player.setMetadata("bw-remove-holo", new FixedMetadataValue(Main.getInstance(), true));
-            player.sendMessage(i18n("click_to_holo_for_remove"));
-        }
-        return true;
+    protected void construct(Command.Builder<CommandSenderWrapper> commandSenderWrapperBuilder) {
+        manager.command(
+                commandSenderWrapperBuilder
+                    .handler(commandContext -> {
+                        // TODO Use Wrapper in the code - Add EyeLocation to PlayerWrapper in ScreamingLib
+                        var player = commandContext.getSender().as(Player.class);
+                        if (!Main.isHologramsEnabled()) {
+                            player.sendMessage(i18n("holo_not_enabled"));
+                        } else {
+                            player.setMetadata("bw-remove-holo", new FixedMetadataValue(Main.getInstance().getPluginDescription().as(JavaPlugin.class), true));
+                            player.sendMessage(i18n("click_to_holo_for_remove"));
+                        }
+                    })
+        );
     }
-
-    @Override
-    public void completeTab(List<String> completion, CommandSender sender, List<String> args) {
-    }
-
 }
