+++ plugin/src/main/java/org/screamingsandals/bedwars/game/Game.java
@@ -41,8 +41,8 @@ import org.screamingsandals.bedwars.api.upgrades.UpgradeStorage;
 import org.screamingsandals.bedwars.api.utils.DelayFactory;
 import org.screamingsandals.bedwars.boss.BossBarSelector;
 import org.screamingsandals.bedwars.boss.XPBar;
-import org.screamingsandals.bedwars.commands.AdminCommand;
 import org.screamingsandals.bedwars.commands.StatsCommand;
+import org.screamingsandals.bedwars.commands.old.AdminCommand;
 import org.screamingsandals.bedwars.config.GameConfigurationContainer;
 import org.screamingsandals.bedwars.config.RecordSave;
 import org.screamingsandals.bedwars.inventories.TeamSelectorInventory;
