+++ plugin/src/main/java/org/screamingsandals/bedwars/inventories/ShopInventory.java
@@ -9,7 +9,7 @@ import org.bukkit.event.Listener;
 import org.bukkit.inventory.ItemStack;
 import org.screamingsandals.bedwars.Main;
 import org.screamingsandals.bedwars.api.events.*;
-import org.screamingsandals.bedwars.commands.DumpCommand;
+import org.screamingsandals.bedwars.commands.old.DumpCommand;
 import org.screamingsandals.bedwars.game.GameStore;
 import org.screamingsandals.bedwars.api.game.ItemSpawnerType;
 import org.screamingsandals.bedwars.api.upgrades.Upgrade;
